+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [roadmap-0.3.0] - 2026-06-16
+
+### Added
+
+- **Ejercicio 03 - Funciones y Alcance**
+- Tests del módulo
+- Documentación del ejercicio
+
+---
+
 ## [roadmap-0.2.0] - 2026-06-16
 
 ### Added